requirements.txt
django==3.0.8
djangorestframework==3.11.0
gunicorn==19.0.0
whitenoise==5.0.1
psycopg2==2.7.7